backend/requirements.txt
anyio==3.6.2
asgiref==3.4.1
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
dataclasses==0.8
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.83.0
h11==0.13.0
httptools==0.3.0
idna==3.4
immutables==0.19
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.31
orjson==3.6.1
protobuf==3.19.6
pycodestyle==2.9.1
pydantic==1.9.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tomli==1.2.3
typing_extensions==4.1.1
ujson==4.0.1
urllib3==1.26.12
uvicorn==0.16.0
uvloop==0.14.0
watchgod==0.7
websockets==9.1
zipp==3.6.0
